ExecutionService の読み取り責務を ExecutionQueryService に分離し、Facade から委譲できるようにする。
