+++ core/application/Statevia.Core.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -24,6 +24,7 @@ public static class ApplicationServiceCollectionExtensions
         services.AddSingleton<IActionSchemaService, ActionSchemaService>();
         services.AddSingleton<IDefinitionSchemaService, DefinitionSchemaService>();
         services.AddSingleton<ExecutionOwnershipTracker>();
+        services.AddScoped<ExecutionQueryService>();
         services.AddScoped<IExecutionService, ExecutionService>();
         services.AddScoped<IEventIngressService, EventIngressService>();
         services.AddScoped<IForkChildExecutionCoordinator, ForkChildExecutionCoordinator>();
